CoreTeam refactoring & exporting fixes. 🔨
