--- src/ignitus-AboutPage/ignitus-About/Components/index.ts
@@ -1,2 +1,0 @@
-
-export { default as About } from './About';
