--- src/ignitus-AboutPage/ignitus-CoreTeam/Components/index.ts
@@ -1,3 +1,0 @@
-
-// eslint-disable-next-line import/prefer-default-export
-export { default as CoreTeam } from './team';
