--- src/ignitus-AboutPage/ignitus-CoreTeam/Containers/TeamContainer.ts
@@ -1,1 +1,0 @@
-export {};
