--- src/ignitus-AboutPage/ignitus-CoreTeam/Containers/index.ts
@@ -1,1 +1,0 @@
-export {};
