--- src/ignitus-AboutPage/ignitus-CoreTeam/actionTypes.ts
@@ -1,1 +1,0 @@
-export {};
