--- src/ignitus-AboutPage/ignitus-CoreTeam/actions.ts
@@ -1,1 +1,0 @@
-export {};
