--- src/ignitus-AboutPage/ignitus-CoreTeam/index.ts
@@ -1,5 +1,0 @@
-import * as components from './Components';
-
-export default {
-  components,
-};
