--- src/ignitus-AboutPage/ignitus-CoreTeam/reducers.ts
@@ -1,1 +1,0 @@
-export {};
