--- src/ignitus-AboutPage/ignitus-CoreTeam/sagas.ts
@@ -1,2 +1,0 @@
-export default function* sagas():IterableIterator<number> {};
-
