--- src/ignitus-AboutPage/ignitus-CoreTeam/selectors.ts
@@ -1,1 +1,0 @@
-export {};
