+++ src/ignitus-Api/index.js
@@ -48,6 +48,7 @@ async function getHeaders(name) {
 }
 
 export async function getContributorsData() {
+  return [];
   try {
     if (navigator.onLine) {
       const frontendConfig = await getHeaders('gh-frontend-contrib');
